+++ decidim-assemblies/spec/shared/manage_assembly_members_examples.rb
@@ -16,28 +16,57 @@ shared_examples "manage assembly members examples" do
     end
   end
 
-  it "creates a new assembly member" do
-    find(".card-title a.new").click
+  context "without existing user" do
+    it "creates a new assembly member" do
+      find(".card-title a.new").click
 
-    execute_script("$('#date_field_assembly_member_designation_date').focus()")
-    find(".datepicker-days .active").click
+      execute_script("$('#date_field_assembly_member_designation_date').focus()")
+      find(".datepicker-days .active").click
 
-    within ".new_assembly_member" do
-      fill_in(
-        :assembly_member_full_name,
-        with: "Daisy O'connor"
-      )
+      within ".new_assembly_member" do
+        fill_in(
+          :assembly_member_full_name,
+          with: "Daisy O'connor"
+        )
 
-      select "President", from: :assembly_member_position
+        select "President", from: :assembly_member_position
 
-      find("*[type=submit]").click
+        find("*[type=submit]").click
+      end
+
+      expect(page).to have_admin_callout("successfully")
+      expect(page).to have_current_path decidim_admin_assemblies.assembly_members_path(assembly)
+
+      within "#assembly_members table" do
+        expect(page).to have_content("Daisy O'connor")
+      end
     end
+  end
 
-    expect(page).to have_admin_callout("successfully")
-    expect(page).to have_current_path decidim_admin_assemblies.assembly_members_path(assembly)
+  context "with existing user" do
+    let!(:member_user) { create :user, organization: assembly.organization }
 
-    within "#assembly_members table" do
-      expect(page).to have_content("Daisy O'connor")
+    it "creates a new assembly member" do
+      find(".card-title a.new").click
+
+      execute_script("$('#date_field_assembly_member_designation_date').focus()")
+      find(".datepicker-days .active").click
+
+      within ".new_assembly_member" do
+        select "Existing user", from: :assembly_member_existing_user
+        autocomplete_select "#{member_user.name} (@#{member_user.nickname})", from: :user_id
+
+        select "President", from: :assembly_member_position
+
+        find("*[type=submit]").click
+      end
+
+      expect(page).to have_admin_callout("successfully")
+      expect(page).to have_current_path decidim_admin_assemblies.assembly_members_path(assembly)
+
+      within "#assembly_members table" do
+        expect(page).to have_content("#{member_user.name} (@#{member_user.nickname})")
+      end
     end
   end
 
@@ -68,7 +97,7 @@ shared_examples "manage assembly members examples" do
       end
     end
 
-    it "deletes a assembly_member" do
+    it "deletes the assembly member" do
       within find("#assembly_members tr", text: assembly_member.full_name) do
         accept_confirm { find("a.action-icon--remove").click }
       end