+++ decidim-participatory_processes/app/cells/decidim/participatory_processes/process_group_cell.rb
@@ -2,9 +2,9 @@
 
 module Decidim
   module ParticipatoryProcesses
-    # This cell renders the process card for an instance of a Process
+    # This cell renders the process card for an instance of a ProcessGroup
     # the default size is the Medium Card (:m)
-    class ProcessCell < Decidim::ViewModel
+    class ProcessGroupCell < Decidim::ViewModel
       def show
         cell card_size, model
       end
@@ -12,7 +12,7 @@ module Decidim
       private
 
       def card_size
-        "decidim/participatory_processes/process_m"
+        "decidim/participatory_processes/process_group_m"
       end
     end
   end